Unify and adapt fuzzing to oss-fuzz interface
