+++ .gitignore
@@ -66,6 +66,7 @@ compile_commands.json
 /autom4te.cache
 /build
 /src/stellar-core
+/src/fuzz_*
 /buckets
 /compile
 /config.h
