+++ Builds/VisualStudio/stellar-core.vcxproj.filters
@@ -94,6 +94,12 @@
     <Filter Include="test">
       <UniqueIdentifier>{632e2a38-59c9-4c6f-8ae3-7e302688207e}</UniqueIdentifier>
     </Filter>
+    <Filter Include="test\fuzz">
+      <UniqueIdentifier>{a1b2c3d4-e5f6-4a5b-9c8d-7e6f5a4b3c2d}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="test\fuzz\targets">
+      <UniqueIdentifier>{b2c3d4e5-f6a7-5b6c-0d9e-8f7a6b5c4d3e}</UniqueIdentifier>
+    </Filter>
     <Filter Include="historyWork">
       <UniqueIdentifier>{eab77361-734d-4f9c-89bd-0a429d67a4d6}</UniqueIdentifier>
     </Filter>
@@ -459,8 +465,26 @@
     <ClCompile Include="..\..\src\work\test\WorkTests.cpp">
       <Filter>work\tests</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\test\fuzz.cpp">
-      <Filter>test</Filter>
+    <ClCompile Include="..\..\src\test\fuzz\FuzzMain.cpp">
+      <Filter>test\fuzz</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\FuzzRegressionTests.cpp">
+      <Filter>test\fuzz</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\FuzzTargetRegistry.cpp">
+      <Filter>test\fuzz</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\FuzzUtils.cpp">
+      <Filter>test\fuzz</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\targets\OverlayFuzzTarget.cpp">
+      <Filter>test\fuzz\targets</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\targets\SorobanFuzzTargets.cpp">
+      <Filter>test\fuzz\targets</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\test\fuzz\targets\TxFuzzTarget.cpp">
+      <Filter>test\fuzz\targets</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\test\TestAccount.cpp">
       <Filter>test</Filter>
@@ -482,9 +506,6 @@
     </ClCompile>
     <ClCompile Include="..\..\src\util\test\BitSetTests.cpp">
       <Filter>util\tests</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\test\FuzzerImpl.cpp">
-      <Filter>test</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\util\FileSystemException.cpp">
       <Filter>util</Filter>
@@ -1683,8 +1704,23 @@
     <ClInclude Include="..\..\src\main\PersistentState.h">
       <Filter>main</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\test\fuzz.h">
-      <Filter>test</Filter>
+    <ClInclude Include="..\..\src\test\fuzz\FuzzTargetRegistry.h">
+      <Filter>test\fuzz</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\test\fuzz\FuzzUtils.h">
+      <Filter>test\fuzz</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\test\fuzz\TxFuzzTarget.h">
+      <Filter>test\fuzz</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\test\fuzz\targets\OverlayFuzzTarget.h">
+      <Filter>test\fuzz\targets</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\test\fuzz\targets\SorobanFuzzTargets.h">
+      <Filter>test\fuzz\targets</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\test\fuzz\targets\TxFuzzTarget.h">
+      <Filter>test\fuzz\targets</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\test\TestAccount.h">
       <Filter>test</Filter>
@@ -1709,9 +1745,6 @@
     </ClInclude>
     <ClInclude Include="..\..\lib\util\cbitset_portability.h">
       <Filter>util</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\test\FuzzerImpl.h">
-      <Filter>test</Filter>
     </ClInclude>
     <ClInclude Include="..\..\lib\util\siphash.h">
       <Filter>lib\util</Filter>
