+++ configure.ac
@@ -330,39 +330,51 @@ $SCCACHE_STATS])
   ])
 ])
 
-# Permit user to enable AFL instrumentation
-AC_ARG_ENABLE([afl],
-              AS_HELP_STRING([--enable-afl],
-                             [build with AFL (fuzzer) instrumentation]))
-AS_IF([test "x$enable_afl" = "xyes"], [
-  AS_IF([test "x$sanitizeopts" != "x"], [
-    AC_MSG_ERROR([AFL is presently incompatible with sanitizers])
+# Permit user to enable libfuzzer-compatible fuzzing (works with libfuzzer,
+# honggfuzz, AFL++, etc.)
+AC_ARG_ENABLE([fuzz],
+              AS_HELP_STRING([--enable-fuzz],
+                             [build with libfuzzer-compatible fuzzing support]))
+AS_IF([test "x$enable_fuzz" = "xyes"], [
+  # Add fuzzing-specific compile-time define. This is the standard define used
+  # by oss-fuzz and enables fuzzing-specific code paths.
+  CXXFLAGS="$CXXFLAGS -DFUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION"
+  CFLAGS="$CFLAGS -DFUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION"
+
+  # LIB_FUZZING_ENGINE must be set to link against a fuzzing engine runtime.
+  # OSS-Fuzz always sets this. For local builds with libfuzzer, you also need
+  # to add instrumentation flags to CXXFLAGS:
+  #
+  #   libfuzzer:
+  #     CXXFLAGS="-O2 -g -fsanitize=fuzzer-no-link" LIB_FUZZING_ENGINE="-fsanitize=fuzzer" \
+  #       ./configure --enable-fuzz
+  #
+  #   honggfuzz:
+  #     CC=hfuzz-clang CXX=hfuzz-clang++ \
+  #       LIB_FUZZING_ENGINE="/path/to/honggfuzz/libhfuzz/libhfuzz.a" \
+  #       ./configure --enable-fuzz
+  #
+  #   AFL++:
+  #     CC=afl-clang-fast CXX=afl-clang-fast++ \
+  #       LIB_FUZZING_ENGINE="/path/to/AFLplusplus/libAFLDriver.a" \
+  #       ./configure --enable-fuzz
+  #
+  # OSS-Fuzz sets CXXFLAGS and LIB_FUZZING_ENGINE automatically.
+  AS_IF([test -z "$LIB_FUZZING_ENGINE"], [
+    AC_MSG_NOTICE([LIB_FUZZING_ENGINE must be set to link against a fuzzing engine runtime.])
+    AC_MSG_NOTICE([For builds with libfuzzer:])
+    AC_MSG_NOTICE([  CXXFLAGS="-O2 -g -fsanitize=fuzzer-no-link" LIB_FUZZING_ENGINE="-fsanitize=fuzzer" ./configure --enable-fuzz])
+    AC_MSG_NOTICE([For honggfuzz:])
+    AC_MSG_NOTICE([  CC=hfuzz-clang CXX=hfuzz-clang++ LIB_FUZZING_ENGINE="/path/to/honggfuzz/libhfuzz/libhfuzz.a" ./configure --enable-fuzz])
+    AC_MSG_NOTICE([For AFL++:])
+    AC_MSG_NOTICE([  CC=afl-clang-fast CXX=afl-clang-fast++ LIB_FUZZING_ENGINE="/path/to/AFLplusplus/libAFLDriver.a" ./configure --enable-fuzz])
+    AC_MSG_NOTICE([OSS-Fuzz sets CXXFLAGS and LIB_FUZZING_ENGINE automatically.])
+    AC_MSG_ERROR([--enable-fuzz requires LIB_FUZZING_ENGINE to be set])
   ])
-  AS_IF([test "x$enable_ccache" = "xyes"], [
-    AC_MSG_ERROR([AFL is presently incompatible with ccache])
-  ])
-  AS_IF([test "x$enable_sccache" = "xyes"], [
-    AC_MSG_ERROR([AFL is presently incompatible with sccache])
-  ])
-  AC_CHECK_PROGS([AFL_FUZZ], [afl-fuzz])
-  AS_CASE(["$CC"],
-          [clang*], [AC_CHECK_PROGS([AFL_CLANG], [afl-clang-fast])
-                     AC_CHECK_PROGS([AFL_CLANGPP], [afl-clang-fast++])
-                     CC="afl-clang-fast"
-                     # below we hard code -std=c++20 since updates to AX_CXX_COMPILE_STDCXX append it to
-                     # CXX, not to CXXFLAGS and thus when setting CXX we override this. For a more detailed explanation
-                     # see: https://github.com/stellar/docker-stellar-core/pull/66#issuecomment-521886881
-                     CXX="afl-clang-fast++ -std=c++20 -DAFL_LLVM_MODE=1"],
-          [gcc*],   [AC_CHECK_PROGS([AFL_GCC], [afl-gcc])
-                     AC_CHECK_PROGS([AFL_GPP], [afl-g++])
-                     CC="afl-gcc"
-                     # below we hard code -std=c++20 since updates to AX_CXX_COMPILE_STDCXX append it to
-                     # CXX, not to CXXFLAGS and thus when setting CXX we override this. For a more detailed explanation
-                     # see: https://github.com/stellar/docker-stellar-core/pull/66#issuecomment-521886881
-                     CXX="afl-g++ -std=c++20"],
-          [AC_MSG_ERROR([Don't know how to instrument CC=$CC with AFL])])
+  AC_MSG_NOTICE([Using LIB_FUZZING_ENGINE=$LIB_FUZZING_ENGINE])
+  AC_SUBST([LIB_FUZZING_ENGINE])
 ])
-AM_CONDITIONAL([USE_AFL_FUZZ], [test "x$enable_afl" == "xyes"])
+AM_CONDITIONAL([ENABLE_FUZZ], [test "x$enable_fuzz" = "xyes"])
 
 # prefer 20 as it's the one we use
 AC_CHECK_PROGS(CLANG_FORMAT, [clang-format-20 clang-format])
@@ -582,7 +594,7 @@ esc() {
 }
 
 # explicitly propagate CFLAGS, CXXFLAGS and LDFLAGS in case they got modified by global options
-ac_configure_args="$ac_configure_args $(esc "CC=$CC" "CXX=$CXX" "CFLAGS=$CFLAGS" "CXXFLAGS=$CXXFLAGS" "LDFLAGS=$LDFLAGS" "RUSTC_WRAPPER=$RUSTC_WRAPPER")"
+ac_configure_args="$ac_configure_args $(esc "CC=$CC" "CXX=$CXX" "CFLAGS=$CFLAGS" "CXXFLAGS=$CXXFLAGS" "LDFLAGS=$LDFLAGS" "RUSTC_WRAPPER=$RUSTC_WRAPPER" "LIB_FUZZING_ENGINE=$LIB_FUZZING_ENGINE")"
 
 AC_CONFIG_HEADERS(config.h)
 AC_CONFIG_FILES(lib/Makefile src/Makefile Makefile)
