+++ make-mks
@@ -31,10 +31,12 @@ message="# This file was generated by make-mks; don't edit it by hand."
 (cd src
  echo "$message"
  echo "SRC_H_FILES" = $(getFiles | egrep -e '\.h$' -e '\.[ih]pp$' | egrep -v '(test|simulation)/' | tr "\n" " ")
- echo "SRC_CXX_FILES" = $(getFiles | egrep '\.cpp$' | egrep -v '(test|simulation)/' | tr "\n" " ")
+ # SRC_CXX_FILES excludes main/main.cpp - it's added explicitly to stellar_core_SOURCES
+ echo "SRC_CXX_FILES" = $(getFiles | egrep '\.cpp$' | egrep -v '(test|simulation)/' | egrep -v '^main/main\.cpp$' | tr "\n" " ")
  echo "SRC_X_FILES" = $(getFiles | egrep '\.x$' | tr "\n" " ")
  echo "SRC_TEST_H_FILES" = $(getFiles | egrep -e '\.h$' -e '\.[ih]pp$' | egrep '(test|simulation)/' | tr "\n" " ")
- echo "SRC_TEST_CXX_FILES" = $(getFiles | egrep '\.cpp$' | egrep '(test|simulation)/' | tr "\n" " ")
+ # SRC_TEST_CXX_FILES excludes FuzzMain.cpp - it's fuzz-target-specific, not part of unit tests
+ echo "SRC_TEST_CXX_FILES" = $(getFiles | egrep '\.cpp$' | egrep '(test|simulation)/' | egrep -v '^test/fuzz/FuzzMain\.cpp$' | tr "\n" " ")
  echo "SRC_RUST_FILES" = $(getFiles | egrep '\.rs$' | tr "\n" " ")
 ) > src/src.mk
 
@@ -58,7 +60,7 @@ listall() {
  echo SOCI_FILES = $(listall soci/src/backends/sqlite3 soci/src/core)
  echo SOCI_PG_FILES = $(listall soci/src/backends/postgresql)
 
- echo SQLITE3_FILES = $(listall sqlite)
+ echo SQLITE3_FILES = $(listall sqlite | grep -v shell)
  echo UTIL_FILES = $(listall util http httpthreaded)
  echo ASIO_H_FILES = $(listall asio/asio/include)
  # The following does not work without boost or -DASIO_STANDALONE=1
