+++ src/test/test.cpp
@@ -565,7 +565,7 @@ getTestConfig(int instanceNumber, Config::TestDbMode mode)
         // seeds the global PRNG might have been seeded with by default (which
         // could thereby collide).
         thisConfig.NODE_SEED = SecretKey::pseudoRandomForTestingFromSeed(
-            0xFFFF0000 + (instanceNumber ^ getLastGlobalStateSeed()));
+            0xFFFF0000 + (instanceNumber ^ getLastGlobalStateSeed().value()));
         thisConfig.NODE_IS_VALIDATOR = true;
 
         // single node setup
